requirements.txt
sanic
pillow
numpy
aiohttp
nose
webtest
